Extract pure env-triad parsers into a stdlib-only module (#481)
The four pure parsing helpers lived in new_env_triad_values_splitter.py,
which imports click, requests, requests_cache, yaml, oaklib and tqdm at
module level. Their unit tests therefore loaded all of that just to test
string handling.

env_triad_parsing.py now holds the three compiled patterns and the four
functions, importing only re and string. The splitter imports them back,
so its public behaviour is unchanged.

Verified: env_triad_parsing pulls in none of click, requests, oaklib,
pymongo, yaml, tqdm or requests_cache. The 7 unit tests run in 0.02s.
Full suite 175 passed. The splitter CLI still resolves --help.
